+++ flask_restplus/swagger.py
@@ -12,7 +12,7 @@ from flask import current_app
 from ._compat import OrderedDict
 
 from . import fields
-from .exceptions import SpecsError
+from .errors import SpecsError
 from .model import ApiModel
 from .utils import merge, not_none, not_none_sorted